+++ src/test/java/domain/GameResultTest.java
@@ -1,30 +1,28 @@
 package domain;
 
-import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import java.util.List;
 
-import java.util.Arrays;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
+import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
 
 class GameResultTest {
-
     @Test
-    @DisplayName("이름으로 결과를 찾을 수 있다.")
-    void findByName() {
-        Player player = new Player(new PlayerName("neo"), new Reward("꽝"));
-        GameResult gameResult = new GameResult(Arrays.asList(player));
+    void 존재하는_이름으로_검색하면_해당_플레이어를_반환한다() {
+        Player pobi = new Player(new PlayerName("pobi"), new Reward("꽝"));
+        GameResult gameResult = new GameResult(List.of(pobi));
 
-        assertThat(gameResult.findByName("neo").getReward().getValue()).isEqualTo("꽝");
+        assertThat(gameResult.findByName("pobi")).isEqualTo(pobi);
     }
 
     @Test
-    @DisplayName("존재하지 않는 이름으로 검색 시 예외가 발생한다.")
-    void findByName_Fail() {
-        GameResult gameResult = new GameResult(Arrays.asList());
+    void 존재하지_않는_이름으로_검색하면_예외가_발생한다() {
+        Player pobi = new Player(new PlayerName("pobi"), new Reward("꽝"));
+        GameResult gameResult = new GameResult(List.of(pobi));
 
-        assertThatThrownBy(() -> gameResult.findByName("none"))
-                .isInstanceOf(IllegalArgumentException.class);
+        assertThatThrownBy(() -> gameResult.findByName("crong"))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessage("존재하지 않는 사람입니다.");
     }
+
 }
