+++ src/test/java/domain/LadderGeneratorTest.java
@@ -1,30 +1,30 @@
 package domain;
 
+import org.assertj.core.api.AssertionsForInterfaceTypes;
 import org.junit.jupiter.api.Test;
-
-import java.util.List;
-
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.*;
+
 
 class LadderGeneratorTest {
     @Test
-    void 존재하는_이름으로_검색하면_해당_플레이어를_반환한다() {
-        Player pobi = new Player(new PlayerName("pobi"), new Reward("꽝"));
-        GameResult gameResult = new GameResult(List.of(pobi));
+    void 빈_공간_발생_시_사다리가_재생성된다() {
+        BooleanGenerator customGenerator = new BooleanGenerator() {
+            private int callCount = 0;
 
-        assertThat(gameResult.findByName("pobi")).isEqualTo(pobi);
-    }
+            @Override
+            public boolean generate() {
+                callCount++;
+                // 폭이 4(간격 3)이고 높이가 2일 때, 1회 생성 시 6번(3 * 2) 호출됨
+                // 첫 6번은 false를 반환하여 모든 간격이 빈 상태를 유도
+                // 두 번째 시도부터 true를 반환하여 정상 생성
+                return callCount > 6;
+            }
+        };
 
-    @Test
-    void 존재하지_않는_이름으로_검색하면_예외가_발생한다() {
-        Player pobi = new Player(new PlayerName("pobi"), new Reward("꽝"));
-        GameResult gameResult = new GameResult(List.of(pobi));
+        Ladder ladder = LadderGenerator.generate(new LadderWidth(4), new LadderHeight(2), customGenerator);
 
-        assertThatThrownBy(() -> gameResult.findByName("crong"))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("존재하지 않는 사람입니다.");
+        AssertionsForInterfaceTypes.assertThat(ladder.getLines()).hasSize(2);
+        assertThat(ladder.getLines().get(0).isConnectedAt(0)).isTrue();
     }
 
 }
